Refactored `rustcoalescence/algorithms/gillespie/src/event_skipping.rs`
